@@ -23,20 +23,20 @@
     "history": "5.3.0",
     "moment": "^2.29.1",
     "qs": "^6.9.4",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.9",
+    "react-dom": "^18.0.9",
     "react-hook-form": "^7.30.0",
     "react-markdown": "^8.0.3",
     "react-router-dom": "6.2.2",
-    "styled-components": "^5.3.3"
+    "styled-components": "^5.3.5"
   },
   "devDependencies": {
     "@types/moment": "^2.13.0",
     "@types/node": "^17.0.21",
     "@types/qs": "^6.9.4",
-    "@types/react": "^17.0.3",
-    "@types/react-dom": "^17.0.3",
-    "@types/styled-components": "5.1.18",
+    "@types/react": "^18.0.9",
+    "@types/react-dom": "^18.0.4",
+    "@types/styled-components": "5.1.25",
     "@typescript-eslint/eslint-plugin": "~4.4.1",
     "@typescript-eslint/parser": "~4.4.1",
     "@vitejs/plugin-react-refresh": "^1.3.2",
@@ -50,7 +50,7 @@
     "eslint-plugin-react-hooks": "^4.1.2",
     "prettier": "~2.1.2",
     "rollup-plugin-copy": "^3.4.0",
-    "typescript": "^4.2.4",
+    "typescript": "4.6.4",
     "vite": "^2.6.14",
     "vite-plugin-checker": "^0.4.5"
   }
