@@ -37,6 +37,6 @@
     "@dfinity/agent": "^0.10.4",
     "@dfinity/candid": "^0.10.4",
     "@dfinity/principal": "^0.10.4",
-    "typescript": "~4.2.4"
+    "typescript": "4.6.4"
   }
 }
