@@ -24,10 +24,10 @@
     "@storybook/builder-vite": "^0.1.33",
     "@storybook/react": "^6.4.22",
     "@storybook/testing-library": "^0.0.11",
-    "@types/react": "^17.0.3",
+    "@types/react": "^18.0.9",
     "@types/react-input-mask": "^3.0.1",
     "@types/react-transition-group": "^4.4.4",
-    "@types/styled-components": "5.1.18",
+    "@types/styled-components": "5.1.25",
     "@typescript-eslint/eslint-plugin": "~4.4.1",
     "@typescript-eslint/parser": "~4.4.1",
     "babel-loader": "^8.2.5",
@@ -46,11 +46,11 @@
     "vite-plugin-checker": "^0.4.5"
   },
   "dependencies": {
-    "react": "^17.0.2",
-    "react-animate-height": "^2.0.23",
+    "react": "^18.0.9",
+    "react-animate-height": "^2.1.2",
     "react-input-mask": "^2.0.4",
     "react-transition-group": "^4.4.2",
-    "styled-components": "^5.3.3",
-    "typescript": "~4.2.4"
+    "styled-components": "^5.3.5",
+    "typescript": "4.6.4"
   }
 }
