@@ -30,6 +30,6 @@
   },
   "dependencies": {
     "@dfinity/candid": "^0.10.4",
-    "typescript": "~4.2.4"
+    "typescript": "4.6.4"
   }
 }
