@@ -1,12 +1,22 @@
 import React from 'react';
 import styled from 'styled-components';
-import { PageWrapper, Pager } from '@union/components';
+import { PageWrapper, Pager, Row as R, Button as B } from '@union/components';
 import { Group } from 'union-ts';
 import { useUnion } from 'services';
+import { NavLink, useParams } from 'react-router-dom';
 import { useCurrentUnion } from '../context';
 import { GroupItem } from './GroupItem';
 
-const Container = styled(PageWrapper)``;
+const Button = styled(B)``;
+const Controls = styled(R)`
+  justify-content: flex-end;
+`;
+
+const Container = styled(PageWrapper)`
+  ${Controls} {
+    margin-bottom: 16px;
+  }
+`;
 
 export interface GroupsProps {
   className?: string;
@@ -14,11 +24,17 @@ export interface GroupsProps {
 }
 
 export const Groups = styled(({ ...p }: GroupsProps) => {
+  const { groupId } = useParams();
   const { principal } = useCurrentUnion();
   const { canister } = useUnion(principal);
 
   return (
     <Container {...p} title='Groups'>
+      <Controls>
+        <Button forwardedAs={NavLink} to='create'>
+          +
+        </Button>
+      </Controls>
       <Pager
         size={5}
         fetch={({ index, size }) =>
@@ -31,7 +47,9 @@ export const Groups = styled(({ ...p }: GroupsProps) => {
             },
           })
         }
-        renderItem={(group: Group) => <GroupItem group={group} />}
+        renderItem={(group: Group) => (
+          <GroupItem group={group} opened={groupId == String(group.id[0])} />
+        )}
       />
     </Container>
   );
