@@ -2,38 +2,36 @@ import React, { useEffect } from 'react';
 import styled from 'styled-components';
 import { PageWrapper, Pager } from '@union/components';
 import { useUnion } from 'services';
-import { AccessConfig } from 'union-ts';
+import { Permission } from 'union-ts';
 import { useCurrentUnion } from '../context';
-import { AccessConfigItem } from './AccessConfigItem';
+import { PermissionItem } from './PermissionItem';
 
 const Container = styled(PageWrapper)``;
 
-export interface AccessConfigsProps {
+export interface PermissionsProps {
   className?: string;
   style?: React.CSSProperties;
 }
 
-export const AccessConfigs = styled(({ ...p }: AccessConfigsProps) => {
+export const Permissions = styled(({ ...p }: PermissionsProps) => {
   const { principal } = useCurrentUnion();
   const { canister } = useUnion(principal);
 
   return (
-    <Container {...p} title='Access configs'>
+    <Container {...p} title='Permissions'>
       <Pager
         size={5}
         fetch={({ index, size }) =>
-          canister.list_access_configs({
+          canister.list_permissions({
             page_req: {
               page_index: index,
               page_size: size,
-              filter: { permission: [], group: [], profile: [] },
               sort: null,
+              filter: { target: [] },
             },
           })
         }
-        renderItem={(item: AccessConfig) => (
-          <AccessConfigItem accessConfig={item} />
-        )}
+        renderItem={(item: Permission) => <PermissionItem permission={item} />}
       />
     </Container>
   );
