@@ -1,7 +1,17 @@
 import React from 'react';
 import styled from 'styled-components';
-import { Text } from '../Text';
+import { Text as T, getFontStyles } from '../Text';
+import { theme } from '../theme';
 import { withBorder } from '../withBorder';
+
+const Text = styled(T)`
+  display: flex;
+
+  &,
+  & > * {
+    ${getFontStyles('p3', 'medium')}
+  }
+`;
 
 const Container = styled.button`
   display: flex;
@@ -9,18 +19,19 @@ const Container = styled.button`
   align-items: center;
 
   background: rgba(0, 0, 0, 0);
-  border: 1px solid grey;
+  border: 1px solid ${({ theme }) => theme.colors.dark};
   padding: 4px 12px;
   text-decoration: none;
   cursor: pointer;
-  color: #373737;
+  color: ${({ theme }) => theme.colors.dark};
+  transition: color 0.2s ease;
 
   &:not([disabled]) {
     transition: background 0.3s ease;
     cursor: pointer;
 
     &:hover {
-      background: rgba(239, 239, 239, 1);
+      color: ${({ theme }) => theme.colors.grey};
     }
   }
 `;
@@ -29,8 +40,13 @@ export interface ButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElemen
 
 export const PureButton = ({ children, ...p }: ButtonProps) => (
   <Container {...p}>
-    <Text variant='p3'>{children}</Text>
+    <Text>{children}</Text>
   </Container>
 );
 
-export const Button = withBorder(PureButton, { withQuad: false, size: 6, color: 'grey' });
+export const Button = withBorder(PureButton, {
+  withQuad: false,
+  size: 6,
+  color: theme.colors.dark,
+  hoverColor: theme.colors.grey,
+});
