@@ -13,13 +13,13 @@ export const Container = styled.div`
 `;
 
 export const Id = styled(Text)<{ len: number }>`
-  color: ${({ theme }) => theme.colors.primary.master.color};
+  color: ${({ theme }) => theme.colors.grey};
   transition: color 200ms ease;
   cursor: pointer;
   min-width: ${({ len }) => len * 8.2 + 10}px;
   padding-top: 1px;
 
   &:hover {
-    color: ${({ theme }) => theme.colors.primary.master.darker};
+    color: ${({ theme }) => theme.colors.dark};
   }
 `;
