@@ -1,15 +1,20 @@
 import React from 'react';
 import styled from 'styled-components';
 import InputMask, { Props as InputMaskProps } from 'react-input-mask';
-import { Text } from '../Text';
+import { Text, getFontStyles } from '../Text';
+import { withBorder } from '../withBorder';
 
 const HelperText = styled(Text)`
   color: red;
 `;
-const Label = styled(Text)``;
-const Input = styled.input`
-  min-height: 32px;
+const Label = styled(Text)`
+  ${getFontStyles('p3', 'medium')}
 `;
+const Input = withBorder(styled.input`
+  min-height: 32px;
+
+  ${getFontStyles('p3', 'regular')}
+`);
 
 const Container = styled.div`
   position: relative;
@@ -37,7 +42,7 @@ export interface TextFieldProps extends React.InputHTMLAttributes<HTMLInputEleme
 export const TextField = React.forwardRef<HTMLInputElement, TextFieldProps>(
   ({ className, style, label, helperText, ...p }, ref) => (
     <Container className={className} style={style}>
-      <Label variant='p1'>{label}</Label>
+      <Label>{label}</Label>
       <Input {...p} ref={ref} />
       {helperText && <HelperText variant='caption'>{helperText}</HelperText>}
     </Container>
@@ -52,7 +57,7 @@ export interface TextAreaProps extends React.TextareaHTMLAttributes<HTMLTextArea
 export const TextArea = React.forwardRef<HTMLTextAreaElement, TextAreaProps>(
   ({ className, style, label, helperText, ...p }, ref) => (
     <Container className={className} style={style}>
-      {label && <Label variant='p1'>{label}</Label>}
+      {label && <Label>{label}</Label>}
       <Input {...p} as='textarea' ref={ref} />
       {helperText && <HelperText variant='caption'>{helperText}</HelperText>}
     </Container>
@@ -67,7 +72,7 @@ export interface MaskedTextFieldProps extends InputMaskProps {
 export const MaskedTextField = React.forwardRef<HTMLInputElement, MaskedTextFieldProps>(
   ({ className, style, label, helperText, ...p }, ref) => (
     <Container className={className} style={style}>
-      <Label variant='p1'>{label}</Label>
+      <Label>{label}</Label>
       <Input {...p} as={InputMask} inputRef={ref} />
       {helperText && <HelperText variant='caption'>{helperText}</HelperText>}
     </Container>
